index.html
I created the starting page of the thing. The page is simple and easy to understand. The section where you enter the code is higher up because it will be used more often. The "Guide" button is red rather than orange so that it stands out more to somebody who's confused. I might change the button colors so that they look less confusing eventually, and add a changing picture at the bottom.