started page nav
There are 3 pages (enterCodePage, mapPage, and guidePage) that are each separate divs in the same html. I added code.js and started on the page navigation. There are now functions for each button and you can navigate between enterCodePage and the beginnings of guidePage, which I've added the heading for using <em>.